+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### Current master
+* Migrate to Swift 5 (#18)
 * Remove client secret from requests (#17)
 * Add prefix to Keychain keys to prevent collision (#15)
 * Return access token in session creation handler (#14)